people table change

+++ migrations/20220327223628_people_main.js
@@ -14,8 +14,8 @@ exports.up = function (knex) {
       table.string("description");
       table.string("address").notNullable();
       table.string("postcode").notNullable();
-      table.integer("latitude");
-      table.integer("longitude");
+      table.float("latitude");
+      table.float("longitude");
       table.timestamp("lastLogin").defaultTo(knex.fn.now());
       table.timestamp("createdAt").defaultTo(knex.fn.now());
     })
